pyproject.toml
[project]
name = "llm-function-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.23.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.64",
]
